requirements.txt
# Backend dependencies
# Install: pip install -r requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
anthropic>=0.42.0
pandas>=2.0.0
python-dotenv>=1.0.0
